Refactoring
Fix imports and change IDistributor to Distributor following the refactoring of file locations into individual folders
